Separate components from LongTerm page - Cycle - Category - Subcategory - Content
